Rearrange files
Files rearranged to disperse code rather than having it all bunched up
