+++ shuttle/request_shuttle/__init__.py
@@ -1,0 +1,49 @@
+from flask import render_template
+from flask_classy import FlaskView, route, request
+from shuttle.db import db_functions as db
+from shuttle.schedules.google_sheets_controller import SheetsController
+from shuttle.schedules.shuttle_schedules_controller import ScheduleController
+from shuttle.shuttle_controller import ShuttleController
+
+
+class RequestShuttleView(FlaskView):
+    def __init__(self):
+        self.sc = ShuttleController()
+        self.ssc = ScheduleController()
+        self.shc = SheetsController()
+
+    @route('/request-shuttle')
+    def request(self):
+        self.sc.check_roles_and_route(['Administrator', 'Driver', 'User'])
+        locations = self.shc.grab_locations()
+        active_requests = db.number_active_requests()
+        if active_requests['requested-pick-up']:
+            position_in_waitlist = db.get_position_in_waitlist()[0]['rownumber']
+        return render_template('request_shuttle/request_shuttle.html', **locals())
+
+    @route('send-request', methods=['GET', 'POST'])
+    def send_shuttle_request_path(self):
+        self.sc.check_roles_and_route(['Administrator', 'Driver', 'User'])
+        json_data = request.get_json()
+        response = db.commit_shuttle_request(json_data['pick-up-location'], json_data['drop-off-location'])
+        if response == 'success':
+            self.sc.set_alert('success', 'Your request has been submitted')
+        elif response == 'same location':
+            self.sc.set_alert('danger', 'Please select two different locations')
+        elif response == 'no location':
+            self.sc.set_alert('danger', 'Please select a location')
+        else:
+            self.sc.set_alert('danger', 'Something went wrong. Please call the ITS Help '
+                                        'Desk at 651-638-6500 for support')
+        return response
+
+    @route('delete-request')
+    def delete_request(self):
+        self.sc.check_roles_and_route(['Administrator', 'Driver', 'User'])
+        request_to_delete = db.delete_current_request()
+        if request_to_delete == 'success':
+            self.sc.set_alert('success', 'Your request has been deleted')
+        else:
+            self.sc.set_alert('danger', 'Something went wrong. Please try again or '
+                                        'call the ITS Help Desk at 651-638-6500')
+        return request_to_delete
