+++ shuttle/templates/driver_check_in/driver_check_in.html
@@ -1,4 +1,4 @@
-{% extends 'shuttle_base.html' %}
+{% extends 'shuttle_bases/shuttle_base.html' %}
 <!-- TODO: select next location after submitting so it is easier for drivers -->
 {% block body_content %}
 <title>Shuttle | {% block page_title %}Driver Check In{% endblock %}</title>
@@ -24,7 +24,7 @@
                 $.ajax({
                     type: 'POST',
                     contentType: 'application/json',
-                    url: "{{ url_for('SchedulesView:load_driver_view') }}",
+                    url: "{{ url_for('DriverCheckInView:load_driver_view') }}",
                     data: JSON.stringify((data)),
                 }).done(function (data) {
                     $('.spinner-img').hide()
