+++ shuttle/templates/driver_check_in/load_driver_check_locations.html
@@ -79,7 +79,7 @@
                 $.ajax({
                     type: 'POST',
                     contentType: 'application/json',
-                    url: "{{ url_for('SchedulesView:send_driver_check_in_info') }}",
+                    url: "{{ url_for('DriverCheckInView:send_driver_check_in_info') }}",
                     data: JSON.stringify((data)),
                 }).done(function (data) {
                     location.reload();
@@ -91,7 +91,7 @@
                 $.ajax({
                     type: 'POST',
                     contentType: 'application/json',
-                    url: "{{ url_for('SchedulesView:send_driver_check_in_info') }}",
+                    url: "{{ url_for('DriverCheckInView:send_driver_check_in_info') }}",
                     data: JSON.stringify((data)),
                 }).done(function (data) {
                     location.reload();
@@ -105,7 +105,7 @@
             $.ajax({
                 type: 'POST',
                 contentType: 'application/json',
-                url: "{{ url_for('SchedulesView:send_driver_break_info') }}",
+                url: "{{ url_for('DriverCheckInView:send_driver_break_info') }}",
                 data: JSON.stringify((data)),
             }).done(function (data) {
                 location.reload()
@@ -117,7 +117,7 @@
             $.ajax({
                 type: 'POST',
                 contentType: 'application/json',
-                url: "{{ url_for('SchedulesView:send_driver_break_info') }}",
+                url: "{{ url_for('DriverCheckInView:send_driver_break_info') }}",
                 data: JSON.stringify((data)),
             }).done(function (data) {
                 location.reload()
