+++ shuttle/templates/schedules/edit_shuttle_schedule.html
@@ -1,4 +1,4 @@
-{% extends 'shuttle_base.html' %}
+{% extends 'shuttle_bases/shuttle_base.html' %}
     <title>Shuttle | {% block page_title %}Edit Schedule{% endblock %}</title>
 {% block body_content %}
 
@@ -10,7 +10,7 @@
         <h6>Please enter locations into the first column. Times should be in the format (00:00 / 0:00) or (-) if the location has no time in a column</h6>
     </div>
     <div class="center-page">
-        {% include 'loaded_views/load_schedule.html' %}
+        {% include 'schedules/load_schedule.html' %}
     </div>
     <div class="center-page" style="padding-top:1em">
         <input id="submit-schedule-button" class="btn btn-warning btn-sm" type="submit" value="Submit to Database">
