+++ shuttle/templates/schedules/shuttle_schedule.html
@@ -1,4 +1,4 @@
-{% extends 'shuttle_base.html' %}
+{% extends 'shuttle_bases/shuttle_base.html' %}
 
 <title>Shuttle | {% block page_title %}Schedule{% endblock %}</title>
 
@@ -8,7 +8,7 @@
         <h4 class="card-title">Shuttle Schedule</h4>
     </div>
     <div class="center-page">
-        {% include 'loaded_views/load_schedule.html' %}
+        {% include 'schedules/load_schedule.html' %}
     </div>
 
 {% endblock %}
